hydrotechDjango_core/backend/requirements.txt
asgiref==3.11.1
Django==4.2.29
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.12.1
sqlparse==0.5.5
typing_extensions==4.15.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
